backend/requirements.txt
beautifulsoup4>=4.12.0
fastapi>=0.115.0
fpdf2>=2.8.0
openai-agents>=0.0.19
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn[standard]>=0.32.0
